folder structured created , added helmet on the server
